Add a new plugin "view-serviceaccount-kubeconfig"
This kubectl plugin shows a kubeconfig to access the apiserver with a
specified serviceaccount.

```
$ kubectl plugin view-serviceaccount-kubeconfig -h
Show a kubeconfig setting to access the apiserver with a specified serviceaccount.

You can also use "view-sa-kubeconfig" as alias for this plugin.

Examples:
  # Show a kubeconfig setting of serviceaccount/default
  kubectl plugin view-serviceaccount-kubeconfig default

Usage:
  kubectl plugin view-serviceaccount-kubeconfig [flags] [options]

Use "kubectl options" for a list of global command-line options (applies to all commands).
```

- https://github.com/superbrothers/kubectl-view-serviceaccount-kubeconfig-plugin