Fix manifest validation in GitHub actions pipeline
Previously, the changed manifests were determined by a git command
equivalent to `git diff <commit>`. This checks for dirty files in the
working directory, so this never detects any changes in the CI where
everything is clean.

Instead, we need to run a command `git diff origin/master <commit>` to
detect all changes that are new in the current branch and are not
present in ref origin/master. For that to work, the master branch needs
to be pulled explicitly.