add kubectl node-pod plugin
This plugin shows the assignment of pods, containers, and jobs to nodes.  

I haven't found something else to do this, and it's been a handy tool for having a node-centric view of what's going on in my k8s clusters.

If this isn't a good fit for a krew plugin, I'm open to suggestions for how to address that, as well as suggestions on how to make this more useful!