Upgrade HNS to v0.5.3 final
Added more documentation as requested in #828 and upgrade to the final
version of the v0.5.3 plugin (from v0.5.3-rc1, which was mainly for
testing purposes).

Tested: ran `kubectl krew install --manifest=plugins/hns.yaml` and got a
nice-looking "caveats" message. I'm not sure how to test the
"description" but it's wrapped now so I'm hoping it looks good.