bump cert-manager plugin versions for 1.8 (#2145)
This process isn't automated because the automation that's provided for
the krew-index isn't a great fit for the cert-manager release process.

Even if it was automated, this PR would require manual review in any
case since we changed our repo name, updated Windows to use zip and
added the .exe extension for the Windows binary.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>