Add nodepools plugin to index (#2853)
* Add nodepools plugin to index

This is in response of #2845 where it was stated that the first
submission shouldn't be automated. This plugin allows users working
with Kubernetes clusters in CSP that have the concept of node
pools/groups to list them in an easier way or listing all the nodes in
a given node pool/group.

Signed-off-by: Leandro López (inkel) <leandro.lopez@grafana.com>

* Remove redundant wording in shortDescription

Signed-off-by: Leandro López (inkel) <leandro.lopez@grafana.com>

* Update plugin description

Signed-off-by: Leandro López (inkel) <leandro.lopez@grafana.com>

* Fix windows binary name

Signed-off-by: Leandro López (inkel) <leandro.lopez@grafana.com>

Signed-off-by: Leandro López (inkel) <leandro.lopez@grafana.com>